Revert "build: Enable builds under LLVM 11. (#172)" (#174)
This reverts commit 5520cf3dff14b3805fafafbfc05853a03aeafcfc.

+++ libs/MII/include/MII.h
@@ -55,9 +55,9 @@
 
 #ifdef CXX_SIMULATOR
 #define volatile
-typedef uintptr_t   mii_reg_t;
+typedef uint64_t    mii_reg_t;
 #else
-typedef uint32_t    mii_reg_t;
+typedef uint16_t    mii_reg_t;
 #endif
 
 /**
